1、添加DSP库，增加中值滤波和iir低通滤波；2、添加注释，规范代码；3、删除多余BSP和不使用的HAL驱动；

+++ bsp/stm32f429-fiberfence/drivers/lcd.h
@@ -3,8 +3,8 @@
 #include "stm32f4xx_hal.h"
 #include "delay.h"
 void lcd1602_init(void);
-void LcdDataWrite(int value);
-void LcdCommandWrite(int value);
+void LcdDataWrite(uint8_t value);
+void LcdCommandWrite(uint8_t value);
 void lcdStrWrite(char *s);
 void lcdDecimalWrite(int value);
 #endif
